Triangle implemented Test split in different file Excluded rule restored in test-ruleset.xml
